+++ stroom-core-client-widget/src/main/java/stroom/widget/dropdowntree/client/view/QuickFilterTooltipUtil.java
@@ -14,19 +14,24 @@ import java.util.stream.Collectors;
 
 public class QuickFilterTooltipUtil {
 
-    public static SafeHtml createTooltip(final String header) {
-        return createTooltip(header, Collections.emptyList());
+    protected static final String FINDING_THINGS_HELP_PAGE = "/user-guide/finding-things/finding-things.html";
+
+    public static SafeHtml createTooltip(final String header,
+                                         final String helpUrlBase) {
+        return createTooltip(header, Collections.emptyList(), helpUrlBase);
     }
 
     public static SafeHtml createTooltip(final String header,
-                                         final List<FilterFieldDefinition> fieldDefinitions) {
+                                         final List<FilterFieldDefinition> fieldDefinitions,
+                                         final String helpUrlBase) {
 
-        return createTooltip(header, null, fieldDefinitions);
+        return createTooltip(header, null, fieldDefinitions, helpUrlBase);
     }
 
     public static SafeHtml createTooltip(final String header,
                                          final Consumer<Builder> preambleBuilder,
-                                         final List<FilterFieldDefinition> fieldDefinitions) {
+                                         final List<FilterFieldDefinition> fieldDefinitions,
+                                         final String helpUrlBase) {
 
         final String defaultFieldNames = getDefaultFieldNamesInfo(fieldDefinitions);
 
@@ -84,7 +89,7 @@ public class QuickFilterTooltipUtil {
                 .addBreak()
                 .appendWithoutBreak("For more information see the ")
                 .appendLinkWithoutBreak(
-                        "https://gchq.github.io/stroom-docs/user-guide/finding-things/finding-things.html",
+                        helpUrlBase + FINDING_THINGS_HELP_PAGE,
                         "Help Documentation")
                 .appendWithoutBreak(".");